signin and accounts
aight, signin and accounts page is done cept for the picture part, I'm
not sure how to store files on the server. right now the only admin
account in the db to test signin with is gk@utd.com pw:123.  also the
only input validation I'm doing right now when adding a user to the db
is making sure the user doesn't already exist and that the fields aren't
blank. I haven't implemented any strict email or password rules.